Make number of weeks in "days view" fluid (#1)
* Make number of weeks in days view fluid, fixes #279

Currently, the number of weeks in "days view" is fixed to 5 which
causes most months to either have a whole week from the previous
month or from the following month being rendered. This is unusual
in most calendar UIs.
This change checks if the previous and current months end on the
last day of the week (depends on the locale set in momentJS) to
decide if an extra week is necessary.
This also fixes tests to take this change into account.

* fixes off-by-one