Enforce tab indentation
It's very common that the indentation is incorrect when someone
publishes a PR, because we're all use to different things and our
editors doesn't always tell us the truth (atleast not IntelliJ). So by
having this rule us maintainers does not need to update the code after a
PR just to fix the indentation - it's already been forced upon the
creator of the PR. This saves us a lot of time.