Migrate to supporting React 16
Some background: I wanted to make a change and I could not get Travis to
run the tests without failing. I did some research and it turns out the
test util package enzyme made some major changes which broke our stuff.
At least that's my theory. I did some testing and the only solution I
found was supporting React from 16.0.0 and above.

Before we supported React from 0.13 and above, and this commit breaks
that backward compability. This is unfortunate but I don't know any
other way of resolving this. I had to exclude some tests because of the
changes for enzyme but I'll take a look at them in the future.