Open calendar when clicking on input element
Before it would just open the calendar on focus, and this would cause
problems for users when they close the calendar and want to open it
again. Before this change the user would have to trigger an onBlur by
clicking somewhere else before triggering onFocus again to open the
calendar again.