Make allowed earliest day selectable in DaysView
If `isValidDate` is provided to set a range of allowed dates, the earliest day of the
range was not selectable in DaysView because it was set to the start of the day. User
would have to implement certain logic in `isValidDate` to make it selectable. This
change would make the earliest day selectable by default.